renamed details directory
