+++ src/pages/public_registration/details-panel/DetailsPanel.tsx
