+++ src/pages/public_registration/details-panel/_components/citation-box/CitationBox.tsx
