+++ src/pages/public_registration/details-panel/_components/citation-box/_utils/format-apa.test.ts
