+++ src/pages/public_registration/details-panel/_components/point-of-contact/PointOfContact.tsx
