+++ src/pages/public_registration/details-panel/_components/point-of-contact/_components/InstitutionsServiceCenterOverview.tsx
